# Speed up recurrent layers by seperating broadcasting of tanh/sigmoid from other parts