```
julia> using Flux, Random, BenchmarkTools

julia> using Flux: GRUCell, gate

julia> model = GRUCell(100, 100);

julia> h = x = randn(Float32, 100, 128);

julia> function ptanh(x::Real)
           l = oftype(x / 1, 1)
           o = oftype(x / 1, 0)
           xθ = oftype(x / 1, 1.92033)
           yθ = oftype(x / 1, 0.96016)
           λ = oftype(x / 1, 0.26037)
           ifelse(x > xθ, yθ, ifelse(x > o, yθ - λ * (x - xθ)^2, ifelse(x > -xθ, λ * (x + xθ)^2 - yθ, -yθ)))
       end
ptanh (generic function with 1 method)

julia> function pσ(x::Real)
           l = oftype(x / 1, 1)
           o = oftype(x / 1, 0)
           h = oftype(x / 1, 0.5)
           x = x / oftype(x / 1, 4.1)
           ifelse(x > l, l, ifelse(x < -l, o, h + x * (l - h * abs(x))))
       end
pσ (generic function with 1 method)

julia> function (m::GRUCell)(h, x)
           b, o = m.b, size(h, 1)
           gx, gh = m.Wi*x, m.Wh*h
           r = σ.(gate(gx, o, 1) .+ gate(gh, o, 1) .+ gate(b, o, 1))
           z = σ.(gate(gx, o, 2) .+ gate(gh, o, 2) .+ gate(b, o, 2))
           h̃ = tanh.(gate(gx, o, 3) .+ r .* gate(gh, o, 3) .+ gate(b, o, 3))
           h′ = (1 .- z).*h̃ .+ z.*h
           return h′, h′
       end


julia> @btime model($h, $x);
  1.262 ms (42 allocations: 801.44 KiB)

julia> function (m::GRUCell)(h, x)
           b, o = m.b, size(h, 1)
           gx, gh = m.Wi*x, m.Wh*h
           r = σ.((r_ = gate(gx, o, 1) .+ gate(gh, o, 1) .+ gate(b, o, 1);))
           z = σ.((z_ = gate(gx, o, 2) .+ gate(gh, o, 2) .+ gate(b, o, 2);))
           h̃ = tanh.((h_ = gate(gx, o, 3) .+ r .* gate(gh, o, 3) .+ gate(b, o, 3);))
           h′ = (1 .- z).*h̃ .+ z.*h
           return h′, h′
       end

julia> @btime model($h, $x);
  1.146 ms (34 allocations: 951.34 KiB)

julia> function (m::GRUCell)(h, x)
           b, o = m.b, size(h, 1)
           gx, gh = m.Wi*x, m.Wh*h
           r = pσ.(gate(gx, o, 1) .+ gate(gh, o, 1) .+ gate(b, o, 1))
           z = pσ.(gate(gx, o, 2) .+ gate(gh, o, 2) .+ gate(b, o, 2))
           h̃ = ptanh.(gate(gx, o, 3) .+ r .* gate(gh, o, 3) .+ gate(b, o, 3))
           h′ = (1 .- z).*h̃ .+ z.*h
           return h′, h′
       end

julia> @btime model($h, $x);
  433.343 μs (42 allocations: 801.44 KiB)

julia> function (m::GRUCell)(h, x)
           b, o = m.b, size(h, 1)
           gx, gh = m.Wi*x, m.Wh*h
           r = pσ.((r_ = gate(gx, o, 1) .+ gate(gh, o, 1) .+ gate(b, o, 1);))
           z = pσ.((z_ = gate(gx, o, 2) .+ gate(gh, o, 2) .+ gate(b, o, 2);))
           h̃ = ptanh.((h_ = gate(gx, o, 3) .+ r .* gate(gh, o, 3) .+ gate(b, o, 3);))
           h′ = (1 .- z).*h̃ .+ z.*h
           return h′, h′
       end

julia> @btime model($h, $x);
  357.087 μs (34 allocations: 951.34 KiB)
```