@CarloLucibello 2020-05-08
That 2.5x speedup given by avx is something worth considering, I would be in favor, but without it the broadcast separation doesn't seem really worth doing
@dhairyagandhi96 @MikeInnes