@AStupidBear 2020-05-08
Later we need to overload broadcasting, which is easier than using `@avx` directly in the cell main body. See https://github.com/chriselrod/LoopVectorization.jl/issues/108