@AStupidBear 2020-05-10
The PR will be useful after https://github.com/FluxML/NNlib.jl/pull/199 is merged