@MikeInnes 2020-05-11
Unfortunately this is type-piratical. I'm not sure that we should try to undo Julia's 'optimisations' here even though they are annoying, because (a) the type piracy could cause issues, (b) it won't help when we do AD (we split broadcasts then anyway) and (c) where performance matters we'll probably rely on ML-specific compiler passes which do this anyway.

It would be nice to have but I think this should be considered a Base Julia issue; we'll benefit automatically once Base does something like `LoopVectorization` by default.