@AStupidBear 2020-07-08
Now that FluxML/NNlib.jl#199 is merged, this PR is useful without type-piracy.