@AStupidBear 2020-10-02
Close in favor of https://github.com/AStupidBear/FastRecurFlux.jl